Finally getting a proper location for the contour related classes instead of mixing them with jsonmap...
