+++ contour/CMakeLists.txt
@@ -1,0 +1,5 @@
+
+add_library(contour STATIC contour.h contour.cxx edge.h interval.h polygon.cxx polygon.h segmentTree.h segmentTree.cxx vertex.h)
+
+add_subdirectory(test)
+
