+++ contour/edge.h
