+++ contour/interval.h
