+++ contour/polygon.cxx
