+++ contour/segmentTree.cxx
