+++ contour/segmentTree.h
