+++ contour/test/testSegmentTree.cxx
@@ -41,8 +41,8 @@ struct YPOS
     YPOS()
     {
 
-      Node* left = new Node{Interval{0,4}};
-      Node* right = new Node{Interval{4,8}};
+      Node* left = new Node{Interval{0, 4}};
+      Node* right = new Node{Interval{4, 8}};
 
       testNode.setLeft(left).setRight(right);
 
@@ -60,30 +60,12 @@ struct YPOS
 
 BOOST_AUTO_TEST_SUITE(o2_mch_geometry)
 
-BOOST_FIXTURE_TEST_SUITE(segment_tree, YPOS)
-
-BOOST_AUTO_TEST_CASE(IntervalCtorThrowsIfBeginIsAfterEnd)
-{
-  BOOST_CHECK_THROW(Interval a(24, 3), std::invalid_argument);
-}
-
-BOOST_AUTO_TEST_CASE(IntervalCtorThrowsIfBeginEqualsEnd)
-{
-  BOOST_CHECK_THROW(Interval a(24, 24), std::invalid_argument);
-}
+BOOST_FIXTURE_TEST_SUITE(segmenttree, YPOS)
 
 BOOST_AUTO_TEST_CASE(NeedAtLeastTwoValuesToBuildASegmentTree)
 {
   BOOST_CHECK_THROW(createSegmentTree({1}), std::invalid_argument);
 }
-
-BOOST_AUTO_TEST_CASE(IntervalIsFullyContainedInInterval)
-{
-  Interval i{1, 5};
-  BOOST_CHECK_EQUAL(Interval(0, 4).isFullyContainedIn(i), false);
-  BOOST_CHECK_EQUAL(Interval(1, 2).isFullyContainedIn(i), true);
-}
-
 
 BOOST_AUTO_TEST_CASE(NodeInsertAndDelete)
 {
@@ -91,12 +73,32 @@ BOOST_AUTO_TEST_CASE(NodeInsertAndDelete)
 
   t->insertInterval(Interval{1, 5});
   t->insertInterval(Interval{5, 8});
-
-  std::cout << (*t) << '\n';
-
   t->deleteInterval(Interval{6, 7});
 
-  std::cout << (*t) << '\n';
+  std::ostringstream os;
+
+  os << '\n' << (*t);
+
+  std::string expectedOutput =
+    R"(
+[0,8] potent
+     [0,4] potent
+           [0,2] potent
+                 [0,1]
+                 [1,2] C=1
+           [2,4] C=1
+                 [2,3]
+                 [3,4]
+     [4,8] potent
+           [4,6] C=1
+                 [4,5]
+                 [5,6]
+           [6,8] potent
+                 [6,7]
+                 [7,8] C=1
+)";
+
+  BOOST_CHECK_EQUAL(os.str(), expectedOutput);
 }
 
 BOOST_AUTO_TEST_CASE(JustCreatedNodeIsNotPotent)
@@ -127,13 +129,6 @@ BOOST_AUTO_TEST_CASE(DemoteNode)
   BOOST_CHECK_EQUAL(testNode.right()->cardinality(), dummyCardinality);
   BOOST_CHECK_EQUAL(testNode.isPotent(), true);
 }
-
-//BOOST_AUTO_TEST_CASE(BuildSegmentTree)
-//{
-//  SegmentTree t{ypos};
-//  std::cout << t << std::endl;
-//  BOOST_CHECK(true);
-//}
 
 BOOST_AUTO_TEST_SUITE_END()
 BOOST_AUTO_TEST_SUITE_END()
