+++ contour/vertex.h
