+++ jsonmap/creator/test/CMakeLists.txt
@@ -4,7 +4,8 @@ include_directories(${CMAKE_SOURCE_DIR}/jsonmap/creator ${BOOST_INCLUDE_DIRS} ${
 
 add_library(mapping SHARED mapping.cxx mapping.h segnumbers.h segnumbers.cxx pcbnumbers.h)
 
-add_executable(testMapping testMapping.cxx testPCB testPads testMotifType.cxx testMotif.cxx testPadSize.cxx testSegmentations.cxx testMotifPosition.cxx testContour.cxx svg.h testSegmentTree.cxx)
+add_executable(testMapping testMapping.cxx testPCB testPads testMotifType.cxx testMotif.cxx testPadSize.cxx testSegmentations.cxx
+        testMotifPosition.cxx)
 
 target_link_libraries(mapping PUBLIC mchjsoncreator Boost::unit_test_framework)
